chore: Update Cilium CLI and Cilium values.yaml
Update the Cilium CLI version to 0.16.11 in the devcontainer and make changes to the Cilium values.yaml file. This includes updating the IPAM mode to "kubernetes" and modifying the security context capabilities for the Cilium agent. Additionally, disable the autoMount feature in the cgroup configuration and set the hostRoot to "/sys/fs/cgroup".